+++ tests/Unit/PaymentTest.php
@@ -4,8 +4,8 @@ namespace Laravel\Cashier\Tests\Unit;
 
 use Laravel\Cashier\Payment;
 use PHPUnit\Framework\TestCase;
-use Stripe\PaymentIntent;
-use Stripe\Subscription;
+use Stripe\PaymentIntent as StripePaymentIntent;
+use Stripe\Subscription as StripeSubscription;
 
 class PaymentTest extends TestCase
 {
@@ -30,7 +30,7 @@ class PaymentTest extends TestCase
     public function test_it_can_return_its_canceled_status()
     {
         $paymentIntent = new PaymentIntent();
-        $paymentIntent->status = Subscription::STATUS_CANCELED;
+        $paymentIntent->status = StripeSubscription::STATUS_CANCELED;
         $payment = new Payment($paymentIntent);
 
         $this->assertTrue($payment->isCanceled());
@@ -43,5 +43,27 @@ class PaymentTest extends TestCase
         $payment = new Payment($paymentIntent);
 
         $this->assertTrue($payment->isSucceeded());
+    }
+
+    public function test_method_calls_are_forward_to_the_stripe_object()
+    {
+        $payment = new Payment(new PaymentIntent());
+
+        $this->assertTrue($payment->cancel()->cancelled);
+    }
+}
+
+class PaymentIntent extends StripePaymentIntent
+{
+    public $cancelled = false;
+
+    /**
+     * @inheritDoc
+     */
+    public function cancel($params = null, $opts = null)
+    {
+        $this->cancelled = true;
+
+        return $this;
     }
 }